Add point and vector creation in tuple4 module
The tuple4 module has been extended to include the creation of points and vectors. This introduces two new functions, point() and vector(), that return Tuple4 structures with values for x, y, z, and (w) which is either 1 for points or 0 for vectors. These additions have been reflected in the raytracer tests with corresponding givens and then statements.